chore: converts gov-notify to typescript
