+++ apps/manage/src/app/service.js
@@ -4,7 +4,7 @@ import { buildInitSharePointDrive } from '#util/sharepoint.js';
 import { MapCache } from '@pins/crowndev-lib/util/map-cache.js';
 import { buildInitEntraClient } from '@pins/crowndev-lib/graph/cached-entra-client.js';
 import { initLogger } from '@pins/crowndev-lib/util/logger.ts';
-import { initGovNotify } from '@pins/crowndev-lib/govnotify/index.js';
+import { initGovNotify } from '@pins/crowndev-lib/govnotify/index.ts';
 import { buildAuditService } from './audit/index.ts';
 import { TextAnalyticsClient } from '@azure/ai-text-analytics';
 import { DefaultAzureCredential } from '@azure/identity';
