+++ apps/portal/src/app/service.js
@@ -5,7 +5,7 @@ import { Client } from '@microsoft/microsoft-graph-client';
 import { DefaultAzureCredential } from '@azure/identity';
 import { TokenCredentialAuthenticationProvider } from '@microsoft/microsoft-graph-client/authProviders/azureTokenCredentials/index.js';
 import { SharePointDrive } from '@pins/crowndev-sharepoint/src/sharepoint/drives/drives.js';
-import { initGovNotify } from '@pins/crowndev-lib/govnotify/index.js';
+import { initGovNotify } from '@pins/crowndev-lib/govnotify/index.ts';
 
 /**
  * This class encapsulates all the services and clients for the application
